integrating select to react-hook-form

+++ src/components/ProductCrudCard/index.tsx
@@ -22,6 +22,12 @@ type Props = {
 
 const ProductCrudCard = ({ product, onDelete }: Props) => {
   const handleDelete = (productId: number) => {
+
+    if(!window.confirm('Tem certeza que deseja deletar?')) {
+      return;
+    }
+
+
     const params: AxiosRequestConfig = {
       method: 'DELETE',
       url: `/products/${productId}`,
